+++ test/chunking-form/samples/preserve-modules-commonjs/_expected/cjs/other.js
@@ -1,0 +1,10 @@
+'use strict';
+
+var value = 43;
+
+var other = {
+	value: value
+};
+
+exports.__moduleExports = other;
+exports.value = value;