+++ test/chunking-form/samples/preserve-modules-commonjs/commonjs.js
@@ -1,0 +1,6 @@
+const external = require('external');
+const { value } = require('./other');
+
+console.log(external, value);
+
+module.exports = 42;