+++ typings/declarations.d.ts
@@ -16,7 +16,6 @@ declare module 'locate-character';
 declare module 'is-reference';
 declare module 'sourcemap-codec';
 declare module 'require-relative';
-declare module 'rollup-pluginutils/src/createFilter.js';
 declare module 'acorn-dynamic-import';
 declare module 'acorn-import-meta';
 declare module 'acorn-bigint';